+++ Configuration/TSconfig/GridElements.ts
@@ -1,4 +1,3 @@
 # === Includes ===
 
-<INCLUDE_TYPOSCRIPT: source="FILE: ./GridElements/T3vContentAccordion.ts">
-<INCLUDE_TYPOSCRIPT: source="FILE: ./GridElements/T3vContentColumnLayout2Columns.ts">
+<INCLUDE_TYPOSCRIPT: source="FILE: ./GridElements/T3vContent.ts">