+++ Configuration/TSconfig/GridElements/T3vContent.ts
@@ -1,0 +1,4 @@
+# === Includes ===
+
+<INCLUDE_TYPOSCRIPT: source="FILE: ./T3vContent/Accordion.ts">
+<INCLUDE_TYPOSCRIPT: source="FILE: ./T3vContent/ColumnLayout.ts">