+++ Configuration/TSconfig/GridElements/T3vContent/ColumnLayout.ts
@@ -1,0 +1,3 @@
+# === Includes ===
+
+<INCLUDE_TYPOSCRIPT: source="FILE: ./ColumnLayout/2Columns.ts">