+++ Configuration/TypoScript/GridElements/T3vContent/ColumnLayout/constants.ts
@@ -1,0 +1,3 @@
+# === Includes ===
+
+<INCLUDE_TYPOSCRIPT: source="FILE: ./2Columns/constants.ts">