+++ Configuration/TypoScript/GridElements/T3vContent/constants.ts
@@ -1,0 +1,4 @@
+# === Includes ===
+
+<INCLUDE_TYPOSCRIPT: source="FILE: ./Accordion/constants.ts">
+<INCLUDE_TYPOSCRIPT: source="FILE: ./ColumnLayout/constants.ts">