+++ Configuration/TypoScript/GridElements/T3vContent/setup.ts
@@ -1,0 +1,4 @@
+# === Includes ===
+
+<INCLUDE_TYPOSCRIPT: source="FILE: ./Accordion/setup.ts">
+<INCLUDE_TYPOSCRIPT: source="FILE: ./ColumnLayout/setup.ts">