+++ Configuration/TypoScript/GridElements/constants.ts
@@ -1,4 +1,3 @@
 # === Includes ===
 
-<INCLUDE_TYPOSCRIPT: source="FILE: ./T3vContentAccordion/constants.ts">
-<INCLUDE_TYPOSCRIPT: source="FILE: ./T3vContentColumnLayout2Columns/constants.ts">
+<INCLUDE_TYPOSCRIPT: source="FILE: ./T3vContent/constants.ts">