+++ Configuration/TypoScript/GridElements/setup.ts
@@ -1,4 +1,3 @@
 # === Includes ===
 
-<INCLUDE_TYPOSCRIPT: source="FILE: ./T3vContentAccordion/setup.ts">
-<INCLUDE_TYPOSCRIPT: source="FILE: ./T3vContentColumnLayout2Columns/setup.ts">
+<INCLUDE_TYPOSCRIPT: source="FILE: ./T3vContent/setup.ts">